.
skoro dovresno, ostalo da se odradi sign-in i sitno poboljsa prikaz tabele korisnika

+++ blueprints/film_blueprint.py
@@ -17,8 +17,10 @@ def getAllFilmovi():
     cursor = mysql.get_db().cursor()
     cursor.execute("SELECT * FROM film")
     filmovi = cursor.fetchall() 
-
-    return flask.jsonify(filmovi)
+    if filmovi is not None:
+        return flask.jsonify(filmovi) 
+    
+    return("Not found", 404)
 
 
 @film_blueprint.route("/<int:film_id>", methods=["GET"])
@@ -35,38 +37,44 @@ def getOneFilm(film_id):
 @film_blueprint.route("", methods=["POST"])    
 @jwt_required()          
 def dodajFilm():
-    baza = mysql.get_db()
-    cursor = baza.cursor() 
-    cursor.execute("INSERT INTO film(naziv, godina_izlaska, reziser, kratak_opis) VALUES(%(naziv)s, %(godina_izlaska)s, %(reziser)s, %(kratak_opis)s)", flask.request.json)
-    baza.commit()
-
-    return flask.request.json, 201 
+    if get_jwt().get("roles") == "ADMIN": 
+        baza = mysql.get_db()
+        cursor = baza.cursor() 
+        cursor.execute("INSERT INTO film(naziv, godina_izlaska, reziser, kratak_opis) VALUES(%(naziv)s, %(godina_izlaska)s, %(reziser)s, %(kratak_opis)s)", flask.request.json)
+        baza.commit()
+        return flask.request.json, 201 
+    
+    return("Nemate prava za ovaj zahtev!", 403)
 
 
 @film_blueprint.route("/<int:film_id>", methods=["PUT"]) 
 @jwt_required()        
 def izmeniFilm(film_id):
-    film = dict(flask.request.json)
-    film["film_id"] = film_id
+    if get_jwt().get("roles") == "ADMIN": 
+        film = dict(flask.request.json)
+        film["film_id"] = film_id
 
-    baza = mysql.get_db()
-    cursor = baza.cursor() 
-    cursor.execute("UPDATE film SET naziv=%(naziv)s, godina_izlaska=%(godina_izlaska)s, reziser=%(reziser)s, kratak_opis=%(kratak_opis)s WHERE id=%(film_id)s", film)
-    baza.commit()
-    cursor.execute("SELECT * FROM film WHERE id=%s", (film_id, ))
-    film = cursor.fetchone() 
+        baza = mysql.get_db()
+        cursor = baza.cursor() 
+        cursor.execute("UPDATE film SET naziv=%(naziv)s, godina_izlaska=%(godina_izlaska)s, reziser=%(reziser)s, kratak_opis=%(kratak_opis)s WHERE id=%(film_id)s", film)
+        baza.commit()
+        cursor.execute("SELECT * FROM film WHERE id=%s", (film_id, ))
+        film = cursor.fetchone() 
+        return flask.jsonify(film)
 
-    return flask.jsonify(film)
+    return("Nemate prava za ovaj zahtev!", 403)
 
 
 @film_blueprint.route("/<int:film_id>", methods=["DELETE"]) 
 @jwt_required()             
-def izbrisiKorisnika(film_id):
-    baza = mysql.get_db()
-    cursor = baza.cursor() 
-    cursor.execute("DELETE FROM film WHERE id=%s", (film_id, ))
-    baza.commit()
+def izbrisiFilm(film_id):
+    if get_jwt().get("roles") == "ADMIN": 
+        baza = mysql.get_db()
+        cursor = baza.cursor() 
+        cursor.execute("DELETE FROM film WHERE id=%s", (film_id, ))
+        baza.commit()
+        return ""
 
-    return ""
+    return("Nemate prava za ovaj zahtev!", 403)
